+++ scripts/game_mode.c
@@ -75,7 +75,7 @@ int mode_choice()
             mouse_pressed(event);
         }
 
-        for (int i = 0; i < sizeof(buttons)/sizeof(buttons[0]); i++)
+        for (int i = 0; i < (int)(sizeof(buttons)/sizeof(buttons[0])); i++)
         {
             out = button_update(buttons[i]);
             switch (out)
@@ -96,7 +96,7 @@ int mode_choice()
             button_render(buttons[i]);
         }
 
-        for (int i = 0; i < sizeof(labels)/sizeof(labels[0]); i++)
+        for (int i = 0; i < (int)(sizeof(labels)/sizeof(labels[0])); i++)
         {
             label_update(labels[i]);
             label_render(labels[i]);
