+++ scripts/pause_menu.c
@@ -62,7 +62,7 @@ int pause_menu()
             mouse_pressed(event);
         }
 
-        for (int i = 0; i < sizeof(buttons)/sizeof(buttons[0]); i++)
+        for (int i = 0; i < (int)(sizeof(buttons)/sizeof(buttons[0])); i++)
         {
             out = button_update(buttons[i]);
             switch (out)
@@ -85,7 +85,7 @@ int pause_menu()
             button_render(buttons[i]);
         }
 
-        for (int i = 0; i < sizeof(labels)/sizeof(labels[0]); i++)
+        for (int i = 0; i < (int)(sizeof(labels)/sizeof(labels[0])); i++)
         {
             label_update(labels[i]);
             label_render(labels[i]);
